Display a message when reboot/shutdown attempt is made on real hardware
When reboot or shutdown command is issued on real hardware, it leaves user in blocked terminal. With this commit, a "safe to shutdown"-style screen is introduced. Made with orange letters on black background to give nostalgic vibes of old Windows versions. Also, somebody please implement real-hardware rebooting and shutdowns.